Fix: change charset of column user_email to utf8 to prevent index size error
The charset has to be corrected before the index gets added. The code manipulates the upgrade section for version 2.4.99.2. The identical database query is also part of the upgrade to version 20220517.1. I was not able to produce an error when altering the column definition again with and to the same settings. So I have to assume, that the script will not break.