Change: check the database server versions and exit the upgrade in necessary
Because of the index size issue with utf8mb4 in MySQL < 5.7(.7) we need to ensure the database server to fulfill the index size for text columns to be able to store more than 767 Bytes. This was introduced with MySQL 5.7.7 (respective with MariaDB 10.2.2). Cancel the upgrade if this minimum requirement is not met.