Style of contact form (#829)
The contact form was more or less unstyled. To match the style of the contact form to the publish form, almost all classes and IDs were added. Personally, I find the excessive use of <fieldset> unnecessary... Thus, this step may only be a temporary solution until the forum is styled by @Auge.